🪛 Event: change `timestamp` type to `Double`

+++ Sources/Events/Event.swift
@@ -74,7 +74,7 @@ open class Event: AnyEvent {
 //    public let target: Any
     
     /// Returns the time (in milliseconds relative to the epoch) at which the event was created
-    public let timeStamp: Int
+    public let timeStamp: Double
     
     public required init (_ event: JSValue) {
         jsEvent = event
